+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## Added
+- [#24](https://github.com/seznam/slo-exporter/pull/24) DynamicClassifier: Allow use of comments in the CSV files, see [the docs](./docs/modules/dynamic_classifier.md#csv-comments).
+
+## Fixed
+- [#24](https://github.com/seznam/slo-exporter/pull/24) DynamicClassifier: Panic on CSV with unexpected number of fields.
 
 ## [v6.3.0-rc1] 2020-08-04
 ## Changed